+++ tests/spec/axiosSingleHttpClient/schema.ts
@@ -1470,6 +1470,7 @@ export interface ApiConfig<SecurityDataType = unknown> extends Omit<AxiosRequest
     securityData: SecurityDataType | null,
   ) => Promise<AxiosRequestConfig | void> | AxiosRequestConfig | void;
   secure?: boolean;
+  format?: ResponseType;
 }
 
 export enum ContentType {
@@ -1483,10 +1484,12 @@ export class HttpClient<SecurityDataType = unknown> {
   private securityData: SecurityDataType | null = null;
   private securityWorker?: ApiConfig<SecurityDataType>["securityWorker"];
   private secure?: boolean;
+  private format?: ResponseType;
 
-  constructor({ securityWorker, secure, ...axiosConfig }: ApiConfig<SecurityDataType> = {}) {
+  constructor({ securityWorker, secure, format, ...axiosConfig }: ApiConfig<SecurityDataType> = {}) {
     this.instance = axios.create({ ...axiosConfig, baseURL: axiosConfig.baseURL || "https://api.github.com" });
     this.secure = secure;
+    this.format = format;
     this.securityWorker = securityWorker;
   }
 
@@ -1507,12 +1510,12 @@ export class HttpClient<SecurityDataType = unknown> {
     };
   }
 
-  public request = async <T = any, E = any>({
+  public request = async <T = any, _E = any>({
     secure,
     path,
     type,
     query,
-    format = "json",
+    format,
     body,
     ...params
   }: FullRequestParams): Promise<AxiosResponse<T>> => {
@@ -1522,6 +1525,7 @@ export class HttpClient<SecurityDataType = unknown> {
         (await this.securityWorker(this.securityData))) ||
       {};
     const requestParams = this.mergeRequestParams(params, secureParams);
+    const responseFormat = (format && this.format) || void 0;
 
     return this.instance.request({
       ...requestParams,
@@ -1530,7 +1534,7 @@ export class HttpClient<SecurityDataType = unknown> {
         ...(requestParams.headers || {}),
       },
       params: query,
-      responseType: format,
+      responseType: responseFormat,
       data: body,
       url: path,
     });