+++ tests/spec/jsAxios/schema.d.ts
@@ -1724,6 +1724,7 @@ export interface ApiConfig<SecurityDataType = unknown> extends Omit<AxiosRequest
     securityData: SecurityDataType | null,
   ) => Promise<AxiosRequestConfig | void> | AxiosRequestConfig | void;
   secure?: boolean;
+  format?: ResponseType;
 }
 export declare enum ContentType {
   Json = "application/json",
@@ -1735,10 +1736,11 @@ export declare class HttpClient<SecurityDataType = unknown> {
   private securityData;
   private securityWorker?;
   private secure?;
-  constructor({ securityWorker, secure, ...axiosConfig }?: ApiConfig<SecurityDataType>);
+  private format?;
+  constructor({ securityWorker, secure, format, ...axiosConfig }?: ApiConfig<SecurityDataType>);
   setSecurityData: (data: SecurityDataType | null) => void;
   private mergeRequestParams;
-  request: <T = any, E = any>({
+  request: <T = any, _E = any>({
     secure,
     path,
     type,